Add summary statistics
Add an option to print summary statistics instead of raw
connections. When this option is true, cnetstat will print the number
of connections between each (container, remote host, remote port)
tuple. It is on by default, because printing raw connections can
easily make the output too large to read on a machine with lots of
connections open.